mcp_prompt_opt/requirements.txt
mcp>=1.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
openai>=1.40.0
anthropic>=0.34.0
httpx>=0.27.0
pytest>=8.3.0
openai-agents